Autogeneration of reference docs.
Follows method described [here](https://mkdocstrings.github.io/recipes/).

While this documents absolutely _everything_, it would be better to find a way to limit the interfaces included rather
than have to manually keep the reference in step with the code. We can always fall back to manual if that can't be done.